Fix compiler errors

+++ include/vector3.h
@@ -305,7 +305,8 @@ inline bool operator!=(Vector3<T> const& lhs, Vector3<T> const& rhs) {
 
 template <typename T>
 cl_float3 toCL(Vector3<T> const& v) {
-    return cl_float3{v.x, v.y, v.z};
+    return cl_float3{static_cast<cl_float>(v.x), static_cast<cl_float>(v.y),
+                     static_cast<cl_float>(v.z)};
 }
 
 }  // namespace mpcs51045
